+++ lib/src/platform/src/codec.dart
@@ -1,5 +1,6 @@
 import 'package:ably_flutter/src/push_notifications/push_notifications.dart';
 import 'package:ably_flutter/src/push_notifications/src/local_device.dart';
+import 'package:ably_flutter/src/realtime/src/realtime_channel_options.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
