+++ lib/src/platform/src/realtime/publish_queue_item.dart
@@ -1,0 +1,12 @@
+import 'dart:async';
+
+import '../../../message/src/message.dart';
+
+/// An item for used to enqueue a message to be published after an ongoing
+/// authCallback is completed
+class PublishQueueItem {
+  List<Message> messages;
+  final Completer<void> completer;
+
+  PublishQueueItem(this.completer, this.messages);
+}
