+++ lib/src/platform/src/realtime/realtime_channel.dart
@@ -1,6 +1,9 @@
 import 'dart:async';
 import 'dart:collection';
 
+import 'package:ably_flutter/src/realtime/src/realtime_channel_interface.dart';
+import 'package:ably_flutter/src/realtime/src/realtime_channel_options.dart';
+import 'package:ably_flutter/src/realtime/src/realtime_channels_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart';
 import 'package:pedantic/pedantic.dart';
@@ -11,6 +14,7 @@ import '../../../message/message.dart';
 import '../../../push_notifications/push_notifications.dart';
 import '../../../realtime/realtime.dart';
 import '../../platform.dart';
+import 'publish_queue_item.dart';
 
 /// Plugin based implementation of Realtime channel
 class RealtimeChannel extends PlatformObject
@@ -58,7 +62,7 @@ class RealtimeChannel extends PlatformObject
     );
   }
 
-  final _publishQueue = Queue<_PublishQueueItem>();
+  final _publishQueue = Queue<PublishQueueItem>();
   Completer<void>? _authCallbackCompleter;
 
   @override
@@ -71,7 +75,7 @@ class RealtimeChannel extends PlatformObject
     messages ??= [
       if (message == null) Message(name: name, data: data) else message
     ];
-    final queueItem = _PublishQueueItem(Completer<void>(), messages);
+    final queueItem = PublishQueueItem(Completer<void>(), messages);
     _publishQueue.add(queueItem);
     unawaited(_publishInternal());
     return queueItem.completer.future;
@@ -226,13 +230,4 @@ class RealtimePlatformChannels
     super.release(name);
     (realtime as Realtime).invoke(PlatformMethod.releaseRealtimeChannel, name);
   }
-}
-
-/// An item for used to enqueue a message to be published after an ongoing
-/// authCallback is completed
-class _PublishQueueItem {
-  List<Message> messages;
-  final Completer<void> completer;
-
-  _PublishQueueItem(this.completer, this.messages);
 }
