+++ lib/src/realtime/src/realtime.dart
@@ -1,7 +1,7 @@
 import '../../authentication/authentication.dart';
 import '../../common/common.dart';
-import '../../push_notifications/push_notifications.dart';
 import '../realtime.dart';
+import 'realtime_channels_interface.dart';
 
 /// an abstract class for Ably's Realtime client
 ///
